Exception support

+++ src/main/java/com/example/store/conrollers/UserController.java
@@ -1,8 +1,9 @@
 package com.example.store.conrollers;
 
 import com.example.store.entities.UserStore;
+import com.example.store.exceptions.NotFoundUserException;
+import com.example.store.exceptions.UserAreReg;
 import com.example.store.services.UserService;
-import jakarta.websocket.server.PathParam;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -23,9 +24,11 @@ public class UserController {
      * @return ифнормацию о регистрации
      */
     @PostMapping(value="users/reg")
-    @ResponseStatus
     public ResponseEntity<UserStore> regUser(@RequestBody UserStore user){
-        return ResponseEntity.created(URI.create("/post")).body(_service.regUser(user));
+        if(_service.getUser(user.getLogin()) == null){
+            return ResponseEntity.ok().body(_service.regUser(user));
+        }
+        throw new UserAreReg();
     }
 
     /**
@@ -33,10 +36,18 @@ public class UserController {
      * @param login логин пользователя
      * @return ифнормацию о входе
      */
-    @GetMapping(value="users/{login}")
+    @GetMapping(value="users/signin")
     @ResponseStatus
-    public  ResponseEntity<UserStore> getUser(@PathVariable(value = "login") String login){
-        return ResponseEntity.ok().body(_service.getUser(login));
+    public  ResponseEntity<UserStore> getUser(@RequestParam String login,
+                                              @RequestParam String password){
+        UserStore localUser = _service.getUser(login);
+        if (localUser == null) {
+            throw new NotFoundUserException();
+        }
+        if (!localUser.getPassword().equals(password)){
+            throw new NotFoundUserException();
+        }
+        return ResponseEntity.ok().body(localUser);
     }
 
     /**
@@ -44,15 +55,25 @@ public class UserController {
      * @param login логин пользователя
      *
      */
-    @PutMapping(value="user/{login}/take/product/{idProduct}")
+    @PutMapping(value="users/{login}/take/product/{idProduct}")
     @ResponseStatus
-    public void addFavorite(@RequestParam String login, @RequestParam int idProduct){
+    public void addFavorite(@PathVariable(value = "login") String login,
+                            @PathVariable(value="idProduct") int idProduct){
+        UserStore localUser = _service.getUser(login);
+        if (localUser == null) {
+            throw new NotFoundUserException();
+        }
         _service.addFavorite(login, idProduct);
     }
 
-    @DeleteMapping(value="user/{login}/delete/product/{idProduct}")
+    @DeleteMapping(value="users/{login}/delete/product/{idProduct}")
     @ResponseStatus
-    public void deleteFavorite(@RequestParam String login, @RequestParam int idProduct){
+    public void deleteFavorite(@PathVariable(value = "login") String login,
+                               @PathVariable(value="idProduct") int idProduct){
+        UserStore localUser = _service.getUser(login);
+        if (localUser == null) {
+            throw new NotFoundUserException();
+        }
         _service.deleteFavorite(login, idProduct);
     }
 }
